refactor, push to history route changes
