+++ frontend/src/RouteView.js
@@ -8,13 +8,14 @@ import Modal from 'react-bootstrap/Modal';
 import styles from './RouteView.module.css';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faClone } from '@fortawesome/free-solid-svg-icons';
+import { getQueryParameters, handleRouting } from './routingUtils.js';
 
 import Map from './map/Map.js';
 import Route from './route/Route.js';
 import OptimizeButton from './route/OptimizeButton.js';
 import SaveShareButtons from './route/SaveShareButtons.js';
 import TripName from './trip-name/TripName.js';
-import { getQueryParameters } from './parameterUtils.js';
+
 /**
  * Render the route page with list of locations in order and directions on a map between the locations.
  */
@@ -57,24 +58,18 @@ function RouteView({ loggedIn }) {
     // save to back end database
   }
 
-  function onManualPlaceChange() {
+  function onManualPlaceChange(newAttractions) {
+    setAttractions(newAttractions);
     setIsOptimized(false);
     setIsSaved(false);
+    tripObject.selectedAttractions = newAttractions;
+    const url = '?trip=' + encodeURIComponent(JSON.stringify(tripObject));
+    history.push(`/route${url}`);
   }
 
   function copyToClipboard(e) {
     textAreaRef.current.select();
     document.execCommand('copy');
-  }
-
-  /**
-   * Creates url and navigates to /explore?trip=
-   * @param {object} history used to route dom with react
-   */
-  function handleRouting(history) {
-    tripObject.selectedAttractions = attractions;
-    const url = '?trip=' + encodeURIComponent(JSON.stringify(tripObject));
-    history.push(`/explore${url}`);
   }
 
   return (
@@ -113,7 +108,7 @@ function RouteView({ loggedIn }) {
       <Button
         variant="link"
         className={styles.editButton}
-        onClick={() => handleRouting(history)}
+        onClick={() => handleRouting(history, 'explore', tripObject, attractions)}
       >
         Edit Attractions
       </Button>
@@ -127,11 +122,7 @@ function RouteView({ loggedIn }) {
         <Row>
           <Col>
             <Row className={styles.routeListContainer}>
-              <Route
-                places={attractions}
-                setPlaces={setAttractions}
-                onManualPlaceChange={onManualPlaceChange}
-              />
+              <Route places={attractions} onManualPlaceChange={onManualPlaceChange} />
             </Row>
             <Row>
               <Container>
